Organized & moved helper & connectivity methods into a class library.
