+++ RedditCrawler.sln
@@ -7,6 +7,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedditCrawler", "RedditCraw
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "rcConfig", "rcConfig\rcConfig.csproj", "{6BBA51E3-41A1-420A-BA80-93271E907A42}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedditCrawlerTests", "RedditCrawlerTests\RedditCrawlerTests.csproj", "{47090832-40F5-4AF6-AE88-1A860CD49665}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "rcListenLibrary", "rcListenLibrary\rcListenLibrary.csproj", "{E465A0C7-45EF-4214-931B-8C9A8A88049E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{6BBA51E3-41A1-420A-BA80-93271E907A42}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6BBA51E3-41A1-420A-BA80-93271E907A42}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6BBA51E3-41A1-420A-BA80-93271E907A42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47090832-40F5-4AF6-AE88-1A860CD49665}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47090832-40F5-4AF6-AE88-1A860CD49665}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47090832-40F5-4AF6-AE88-1A860CD49665}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47090832-40F5-4AF6-AE88-1A860CD49665}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E465A0C7-45EF-4214-931B-8C9A8A88049E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E465A0C7-45EF-4214-931B-8C9A8A88049E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E465A0C7-45EF-4214-931B-8C9A8A88049E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E465A0C7-45EF-4214-931B-8C9A8A88049E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
