+++ RedditCrawlerTests/rcConfigTests.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using rcConfig;
+using RedditCrawler;
+
+namespace RedditCrawlerTests
+{
+    [TestClass]
+    public class rcConfigTests
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+            
+        }
+    }
+}
