+++ rcConfig/MainWindow.xaml
@@ -6,7 +6,65 @@
         xmlns:local="clr-namespace:rcConfig"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
+
     <Grid>
+        <Grid.Resources>
+            <LinearGradientBrush x:Key="ButtonNormalBackground" EndPoint="0,1" StartPoint="0,0">
+                <GradientStop Color="#F3F3F3" Offset="0"/>
+                <GradientStop Color="#EBEBEB" Offset="0.5"/>
+                <GradientStop Color="#DDDDDD" Offset="0.5"/>
+                <GradientStop Color="#CDCDCD" Offset="1"/>
+            </LinearGradientBrush>
+            <SolidColorBrush x:Key="ButtonNormalBorder" Color="#FF707070"/>
+            <Style x:Key="ButtonFocusVisual">
+                <Setter Property="Control.Template">
+                    <Setter.Value>
+                        <ControlTemplate>
+                            <Rectangle Margin="2" SnapsToDevicePixels="true" Stroke="{DynamicResource {x:Static SystemColors.ControlTextBrushKey}}" StrokeThickness="1" StrokeDashArray="1 2"/>
+                        </ControlTemplate>
+                    </Setter.Value>
+                </Setter>
+            </Style>
+            <Style x:Key="TBStyle1" TargetType="{x:Type ToggleButton}">
+                <Setter Property="FocusVisualStyle" Value="{StaticResource ButtonFocusVisual}"/>
+                <Setter Property="Background" Value="{StaticResource ButtonNormalBackground}"/>
+                <Setter Property="BorderBrush" Value="{StaticResource ButtonNormalBorder}"/>
+                <Setter Property="BorderThickness" Value="1"/>
+                <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.ControlTextBrushKey}}"/>
+                <Setter Property="HorizontalContentAlignment" Value="Center"/>
+                <Setter Property="VerticalContentAlignment" Value="Center"/>
+                <Setter Property="Padding" Value="1"/>
+                <Setter Property="Template">
+                    <Setter.Value>
+                        <ControlTemplate TargetType="{x:Type ToggleButton}">
+                            <StackPanel Orientation="Horizontal">
+                                <ContentPresenter HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}" Margin="{TemplateBinding Padding}" RecognizesAccessKey="True" SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" VerticalAlignment="{TemplateBinding VerticalContentAlignment}"/>
+                                <Border x:Name="on" Width="25" Height="25" Background="LightGray" CornerRadius="2,0,0,4" Margin="10,0,0,0">
+                                    <TextBlock x:Name="onText" Text="On" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+                                </Border>
+                                <Border x:Name="off" Width="25" Height="25" Background="LightGray" CornerRadius="0,2,4,0">
+                                    <TextBlock x:Name="offText" Text="Off" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+                                </Border>
+                            </StackPanel>
+                            <ControlTemplate.Triggers>
+                                <Trigger Property="IsEnabled" Value="false">
+                                    <Setter Property="Foreground" Value="#ADADAD"/>
+                                </Trigger>
+                                <Trigger Property="IsChecked" Value="true">
+                                    <Setter TargetName="on" Property="Background" Value="LightBlue"/>
+                                    <Setter TargetName="onText" Property="Foreground" Value="White"/>
+                                </Trigger>
+                                <Trigger Property="IsChecked" Value="False">
+                                    <Setter TargetName="off" Property="Background" Value="LightBlue"/>
+                                    <Setter TargetName="offText" Property="Foreground" Value="White"/>
+                                </Trigger>
+                            </ControlTemplate.Triggers>
+                        </ControlTemplate>
+                    </Setter.Value>
+                </Setter>
+            </Style>
+            
+        </Grid.Resources>
         <Frame Content="RedditCrawler Parameter Configuration" Background="AliceBlue"  HorizontalAlignment="Left" Height="399" Margin="10,10,0,0" VerticalAlignment="Top" Width="772"/>
         <Label Content="Subreddit to monitor:" FontSize="16" HorizontalAlignment="Left" Margin="10,65,0,0" VerticalAlignment="Top"/>
         <TextBox Name="txtSubreddit" HorizontalAlignment="Left" Height="23" Margin="195,73,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="194"/>
@@ -18,8 +76,8 @@
         <TextBox Name="txtEmail" HorizontalAlignment="Left" Height="23" Margin="195,295,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="194"/>
         <Label Content="Email Password:" FontSize="16" HorizontalAlignment="Left" Margin="10,317,0,0" VerticalAlignment="Top"/>
         <PasswordBox Name="pwdEmail" HorizontalAlignment="Left" Margin="195,325,0,0" VerticalAlignment="Top" Width="194" Height="23"/>
-        <Label Content="Search Terms (separate by line):" FontSize="16" HorizontalAlignment="Left" Margin="417,37,0,0" VerticalAlignment="Top"/>
-        <RichTextBox Name="rtfSearchTerms" HorizontalAlignment="Left" Height="275" Margin="417,73,0,0" VerticalAlignment="Top" Width="283">
+        <Label Content="Search Terms (separate by line):" FontSize="16" HorizontalAlignment="Left" Margin="417,11,0,0" VerticalAlignment="Top"/>
+        <RichTextBox Name="rtfSearchTerms" HorizontalAlignment="Left" Height="275" Margin="417,47,0,0" VerticalAlignment="Top" Width="283">
             <FlowDocument>
                 <Paragraph>
                     <Run Text=""/>
@@ -30,6 +88,8 @@
         <Button x:Name="btnSubmitSub" Content="Submit Subreddit" FontSize="14" HorizontalAlignment="Left" Margin="257,101,0,0" VerticalAlignment="Top" Width="132" Click="btnSubmitSub_click"/>
         <Button x:Name="btnSubmitLogin" Content="Submit Login" FontSize="14" HorizontalAlignment="Left" Margin="257,229,0,0" VerticalAlignment="Top" Width="132" Click="btnSubmitLogin_click"/>
         <Button x:Name="btnSubmitEmail" Content="Submit Email" FontSize="14" HorizontalAlignment="Left" Margin="257,353,0,0" VerticalAlignment="Top" Width="132" Click="btnSubmitEmail_click"/>
-
+        <ToggleButton 
+        Content="ON LINE MODE" 
+            Style="{StaticResource TBStyle1}" Margin="556,139,-556,-139"/>
     </Grid>
 </Window>
