+++ rcListenLibrary/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net461" />
+  <package id="RedditSharp" version="1.1.14" targetFramework="net461" />
+</packages>
